requirements.txt
torch>=2.1
torchvision>=0.16
opencv-python
pyyaml
tqdm
numpy
matplotlib
pycocotools
fastapi
uvicorn
python-multipart
gradio
requests
onnxruntime
